requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.0
PyYAML==6.0.2
